Move Logo List to its own file
